+++ lib/timber/util.rb
@@ -1,7 +1,7 @@
 require "timber/util/active_support_log_subscriber"
+require "timber/util/attribute_normalizer"
 require "timber/util/hash"
-require "timber/util/http_event"
-require "timber/util/object"
+require "timber/util/non_nil_hash_builder"
 require "timber/util/request"
 require "timber/util/struct"